Rename to termination_check module
The runs concept is gone.
